Combine RIO datasets

This function combines multiple RIO datasets based on a common ID field.

Usage

rio_combine(..., by = NULL, suffix = c("_x", "_y"))

Arguments

...

Named tibbles to combine. Each argument should be a tibble with data from RIO.

by

Character vector of columns to join by. If NULL, will join by all columns with the same name.

suffix

Suffixes to add to make non-by column names unique across the joined tables.

Value

A combined tibble with data from all input tibbles.

Examples

if (FALSE) { # \dontrun{
rio_conn <- rio_api_connection()
locations <- rio_get_locations(rio_conn, city = "Rotterdam")
institutions <- rio_fetch_data(rio_conn, resource_id = "institution-resource-id")

# Combine datasets
combined_data <- rio_combine(
  locations = locations,
  institutions = institutions,
  by = "INSTELLINGSCODE"
)
} # }