Fetch data from a RIO dataset

This function retrieves data from a specific dataset in the RIO CKAN API. You can specify either a resource_id or a resource_name to identify the dataset. Additional filter parameters can be provided to filter the data. By default, this function retrieves all available records (not just the first 1000).

Usage

rio_fetch_data(
  conn = NULL,
  resource_id = NULL,
  resource_name = NULL,
  limit = NULL,
  query = NULL,
  package_id = "rio_nfo_po_vo_vavo_mbo_ho",
  all_records = TRUE,
  batch_size = 1000,
  quiet = FALSE,
  ...
)

Arguments

conn

A connection object created with rio_api_connection(). If NULL, a new connection will be created.

resource_id

The ID of the dataset resource to retrieve. Default is NULL.

resource_name

The name of the dataset resource to retrieve. Default is NULL. Either resource_id or resource_name must be provided.

limit

Maximum number of records to return. Default is NULL (all records). Specify a number to limit the results.

query

A search query string for full-text search. Default is NULL.

package_id

The ID of the package to search in when using resource_name. Default is "rio_nfo_po_vo_vavo_mbo_ho".

all_records

Logical indicating whether to fetch all records (which may require multiple API calls). Default is TRUE.

batch_size

Number of records to retrieve per API call. Default is 1000, which is the maximum supported by the API.

quiet

Logical indicating whether to suppress progress messages. Default is FALSE.

...

Additional named parameters used as filters. For example, PLAATSNAAM = "Rotterdam" will filter for records where the PLAATSNAAM field equals "Rotterdam".

Value

A tibble containing the retrieved data.

Examples

if (FALSE) { # \dontrun{
# Get all educational locations in Rotterdam
locations <- rio_fetch_data(
  resource_name = "Onderwijslocaties",
  PLAATSNAAM = "Rotterdam"
)

# Get only the first 500 records
limited_data <- rio_fetch_data(
  resource_name = "Onderwijslocaties",
  limit = 500
)

# Get the first 1000 records without attempting to fetch all records
first_batch <- rio_fetch_data(
  resource_name = "Onderwijslocaties",
  all_records = FALSE
)

# Fetch data without showing progress bar
silent_fetch <- rio_fetch_data(
  resource_name = "Onderwijslocaties",
  quiet = TRUE
)
} # }