Create a map of educational locations

This function creates a leaflet map of educational locations from RIO data.

Usage

rio_map_locations(locations, popup_vars = NULL, color = "blue", radius = 5)

Arguments

locations

A tibble containing educational location data from the RIO API.

popup_vars

Character vector of variable names to include in popups. Default is NULL.

color

The color of the markers. Default is "blue".

radius

The radius of the markers. Default is 5.

Value

A leaflet map object.

Examples

if (FALSE) { # \dontrun{
locations <- rio_get_locations(city = "Rotterdam")
map <- rio_map_locations(
  locations,
  popup_vars = c("ONDERWIJSLOCATIENAAM", "PLAATSNAAM")
)
} # }